requirements.txt
click==8.1.7
eth_abi==5.1.0
python-dotenv==1.0.1
